Moved ProjectReleases and ProjectReleaseLinks to different modules
